updating dev branch

+++ src/ChargeInputModule.rs
@@ -65,14 +65,14 @@ static mut player_6_per_direction_vec: Vec<Vec<per_dir>> = Vec::new();
 static mut player_7_per_direction_vec: Vec<Vec<per_dir>> = Vec::new();
 static mut player_8_per_direction_vec: Vec<Vec<per_dir>> = Vec::new();
 
-static mut player_1_last_frame: i32 = 0;
-static mut player_2_last_frame: i32 = 0;
-static mut player_3_last_frame: i32 = 0;
-static mut player_4_last_frame: i32 = 0;
-static mut player_5_last_frame: i32 = 0;
-static mut player_6_last_frame: i32 = 0;
-static mut player_7_last_frame: i32 = 0;
-static mut player_8_last_frame: i32 = 0;
+static mut player_1_last_frame: f32 = 0.0;
+static mut player_2_last_frame: f32 = 0.0;
+static mut player_3_last_frame: f32 = 0.0;
+static mut player_4_last_frame: f32 = 0.0;
+static mut player_5_last_frame: f32 = 0.0;
+static mut player_6_last_frame: f32 = 0.0;
+static mut player_7_last_frame: f32 = 0.0;
+static mut player_8_last_frame: f32 = 0.0;
 
 /// Adds a charge input to the Moveset
 /// 
@@ -344,7 +344,7 @@ pub unsafe fn reset_module(module_accessor:*mut BattleObjectModuleAccessor) {
 
     *per_dir = Vec::new();
     *per_input = Vec::new();
-    *last_update = 0;
+    *last_update = 0.0;
 
 }
 /// Updates the life of each charge input
@@ -387,7 +387,7 @@ pub unsafe fn set_stick_type(entry_id: usize, input: usize, new_stick_type: Stic
 }
 
 ///updates all the charge inputs for that frame 
-pub unsafe fn update_module(module_accessor:*mut BattleObjectModuleAccessor, frame: i32, ignore_repeat_frame: bool) {
+pub unsafe fn update_module(module_accessor:*mut BattleObjectModuleAccessor, frame: f32, ignore_repeat_frame: bool) {
 
     let entry_id = WorkModule::get_int(module_accessor, *FIGHTER_INSTANCE_WORK_ID_INT_ENTRY_ID) as usize;
     let last_update_frame = get_last_update_frame(&entry_id);
@@ -468,7 +468,7 @@ pub unsafe fn update_module(module_accessor:*mut BattleObjectModuleAccessor, fra
                 else {
 
                     break;
-                    
+
                 }
             }
         }
@@ -965,7 +965,7 @@ unsafe fn get_per_dir_vec(entry_id: &usize) -> &mut Vec<Vec<per_dir>> {
 
     ret
 }
-unsafe fn get_last_update_frame(entry_id: &usize) -> *mut i32 {
+unsafe fn get_last_update_frame(entry_id: &usize) -> *mut f32 {
     let ret = if *entry_id == 0 {&raw mut player_1_last_frame} else if *entry_id == 1 { &raw mut player_2_last_frame }
         else if *entry_id == 2 { &mut player_3_last_frame } else if *entry_id == 3 { &mut player_4_last_frame }
         else if *entry_id == 4 { &mut player_5_last_frame } else if *entry_id == 5 { &mut player_6_last_frame }
